More cleanup * refactor to add props in rather than derive in provider * makes it more flexible
