+++ packages/story-editor/src/app/mediaUpload/provider.js
@@ -16,16 +16,13 @@
 /**
  * External dependencies
  */
-import { canTranscodeResource, resourceList } from '@web-stories-wp/media';
-import { sprintf, translateToExclusiveList, __ } from '@web-stories-wp/i18n';
-import { useCallback, useMemo, useRef } from '@web-stories-wp/react';
-import { useSnackbar } from '@web-stories-wp/design-system';
+
+import { useRef } from '@web-stories-wp/react';
 import PropTypes from 'prop-types';
 /**
  * Internal dependencies
  */
-import { useConfig, useLocalMedia, useStory } from '..';
-import useFFmpeg from '../media/utils/useFFmpeg';
+import { useConfig } from '..';
 import Context from './context';
 
 function OpenMediaUploadHandler({ onOpenMediaUpload, ...props }) {
@@ -87,149 +84,13 @@ OpenMediaUploadHandler.propTypes = {
  */
 function MediaUploadProvider({ children, ...props }) {
   const {
-    allowedTranscodableMimeTypes,
-    allowedFileTypes,
-    allowedMimeTypes: {
-      image: allowedImageMimeTypes,
-      video: allowedVideoMimeTypes,
-    },
     // capabilities: { hasUploadMediaAction },
     MediaUpload,
   } = useConfig();
-  const { showSnackbar } = useSnackbar();
-  const { selectedElements, updateElementsById } = useStory(
-    ({ state: { selectedElements }, actions: { updateElementsById } }) => ({
-      selectedElements,
-      updateElementsById,
-    })
-  );
-  const { resetWithFetch, updateVideoIsMuted, optimizeVideo, optimizeGif } =
-    useLocalMedia(
-      ({
-        actions: {
-          resetWithFetch,
-          updateVideoIsMuted,
-          optimizeVideo,
-          optimizeGif,
-        },
-      }) => {
-        return {
-          resetWithFetch,
-          updateVideoIsMuted,
-          optimizeVideo,
-          optimizeGif,
-        };
-      }
-    );
-
-  const { isTranscodingEnabled } = useFFmpeg();
-
-  const allowedMimeTypes = useMemo(() => {
-    if (isTranscodingEnabled) {
-      return [
-        ...allowedTranscodableMimeTypes,
-        ...allowedImageMimeTypes,
-        ...allowedVideoMimeTypes,
-      ];
-    }
-    return [...allowedImageMimeTypes, ...allowedVideoMimeTypes];
-  }, [
-    allowedImageMimeTypes,
-    allowedVideoMimeTypes,
-    isTranscodingEnabled,
-    allowedTranscodableMimeTypes,
-  ]);
-
-  const transcodableMimeTypes = useMemo(() => {
-    return allowedTranscodableMimeTypes.filter(
-      (x) => !allowedVideoMimeTypes.includes(x)
-    );
-  }, [allowedTranscodableMimeTypes, allowedVideoMimeTypes]);
-
-  let onSelectErrorMessage = __(
-    'No file types are currently supported.',
-    'web-stories'
-  );
-  if (allowedFileTypes.length) {
-    onSelectErrorMessage = sprintf(
-      /* translators: %s: list of allowed file types. */
-      __('Please choose only %s to insert into page.', 'web-stories'),
-      translateToExclusiveList(allowedFileTypes)
-    );
-  }
-
-  /**
-   * Insert element such image, video and audio into the editor.
-   *
-   * @param {Object} resource Resource object
-   * @param {string} thumbnailURL The thumbnail's url
-   * @return {null|*} Return onInsert or null.
-   */
-  const insertMediaElement = useCallback(
-    (resource, thumbnailURL) => {
-      resourceList.set(resource.id, {
-        url: thumbnailURL,
-        type: 'cached',
-      });
-      updateElementsById({
-        elementIds: [selectedElements?.[0]?.id],
-        properties: { resource },
-      });
-    },
-    [selectedElements, updateElementsById]
-  );
-
-  const handleMediaSelect = useCallback(
-    (resource) => {
-      try {
-        if (isTranscodingEnabled && canTranscodeResource(resource)) {
-          if (transcodableMimeTypes.includes(resource.mimeType)) {
-            optimizeVideo({ resource });
-          }
-
-          if (resource.mimeType === 'image/gif') {
-            optimizeGif({ resource });
-          }
-        }
-        // WordPress media picker event, sizes.medium.source_url is the smallest image
-        insertMediaElement(
-          resource,
-          resource.sizes?.medium?.source_url || resource.src
-        );
-
-        if (
-          !resource.local &&
-          allowedVideoMimeTypes.includes(resource.mimeType) &&
-          resource.isMuted === null
-        ) {
-          updateVideoIsMuted(resource.id, resource.src);
-        }
-      } catch (e) {
-        showSnackbar({
-          message: e.message,
-          dismissable: true,
-        });
-      }
-    },
-    [
-      allowedVideoMimeTypes,
-      insertMediaElement,
-      isTranscodingEnabled,
-      optimizeGif,
-      optimizeVideo,
-      showSnackbar,
-      transcodableMimeTypes,
-      updateVideoIsMuted,
-    ]
-  );
 
   return (
     <MediaUpload
       {...props}
-      onSelect={handleMediaSelect}
-      onClose={resetWithFetch}
-      type={allowedMimeTypes}
-      onSelectErrorMessage={onSelectErrorMessage}
       render={(open) => (
         /* the only way to access `open` is through the render prop of `MediaUpload` */
         <OpenMediaUploadHandler onOpenMediaUpload={open}>
@@ -249,7 +110,10 @@ MediaUploadProvider.propTypes = {
   onSelect: PropTypes.func.isRequired,
   onSelectErrorMessage: PropTypes.string.isRequired,
   title: PropTypes.string.isRequired,
-  type: PropTypes.string.isRequired,
+  type: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.arrayOf(PropTypes.string),
+  ]).isRequired,
 };
 
 export default MediaUploadProvider;
