+++ packages/story-editor/src/components/canvas/navLayer.js
@@ -29,7 +29,6 @@ import { useQuickActions } from '../../app/highlights';
 import DirectionAware from '../directionAware';
 import Carousel from '../carousel';
 import { useLayout } from '../../app';
-import { useMediaUpload } from '../../app/mediaUpload';
 import {
   CarouselArea,
   HeadArea,
@@ -42,8 +41,7 @@ function NavLayer({ header }) {
   const { hasHorizontalOverflow } = useLayout(
     ({ state: { hasHorizontalOverflow } }) => ({ hasHorizontalOverflow })
   );
-  const { onOpenMediaUpload } = useMediaUpload();
-  const quickActions = useQuickActions({ onOpenMediaUpload });
+  const quickActions = useQuickActions();
 
   /**
    * Stop the event from bubbling if the user clicks in between buttons.
